Remove diagnostic signs in line number column
Since tiny-inline-diagnostic shows diagnostics to the right of the line
in question, it's redundant (and distracting) to see signs also in the
line number column (on the left), signs which complete for space with
git diff status indicators (line add/remove/update).

In some cases the diagnostic on the right is out of the visible
viewport and gets missed. The code in question does get underlined,
which mitigates this problem, but it's not fool proof.

On ballance I think it's better to simplify what's being communicated,
and optimise for ease of use, and then get that experience fast and
enjoyable. For example <C-Up> and <C-Down> can be used to jump between
diagnostics to quickly check for any out of viewport diagnostics.