Replace "Top   1:3" w/ "1/128:3" (progress/location) in status line
Previous approach used a pre-computed width to prevent width in the
status line as the data changed. The trade off was large gaps in the
text at smaller line numbers. This made sense with the original
visual design of lualine, but my subdued approach uses text position
to draw (or evade) attention.