Add nvim-cmp ctrl+F filter, borders + fix rust-analyzer disabled Leptos macro
When autocompleting fields or method names, it's often distracting to
see numerous text snippets. One option is to de-prioritise text snippets
to always appear later in the list, but I liked the idea of
interactively filtering for exactly what one is interested in when
looking at a variable such as when typing `my_variable.` to see what is
available in the way of fields and methods. Pressing ctrl+F while
inside of the nvim-cmp popup will filter on all such pertenant kinds of
autocomplete entries. Pressing ctrl+F again will revert back to normal.
Closing the popup also resets the filter.

Fixed an undesirable modification whereby I asked rust-analyzer to
ignore a particular Leptos macro called "component". I didn't realise
that it would underline ever instance as an error, not what I wanted, as
I use the Leptos component macro liberally.

Went back to using visual borders on the nvim-cmp popup window. It's
slightly more distracting in normal use, but makes all the difference
when only a few entries are showing, or when the documentation for an
entry is short. When either of the aforementioned happens, it was very
easy to mistake the popup for the code in the buffer. The addition of
the visual border helps distinguish the two more clearly.

I also set nvim-cmp's performance characteristics to have no debouncing
or throttling. I couldn't tell if it made a difference. I think most of
the waiting time is due to the LSP, and Rust, the language I'm currently
using it with, is one of the slower LSP experiences.