Re-add rounded borders for LSP hover, now w/ `:w` bug
Treesitter was complaining whenever I unmapped `K` in the `LspAttach`
callback. Everything was fine on the first buffer opened, but on the
second the text would load in without any colour highlighting, and then
attempting to wright the buffer would throw an error saying saying the
file is already open (or some such), and you have to force write to
overcome it, and highlighting is different from then on, in a way which
makes me think treesitter is falling back to some default highlighting
in lieu of LSP input.